+++ pkg/dds/test/dap/integration/debug_breakpoints_test.dart
@@ -4,6 +4,7 @@
 
 import 'dart:io';
 
+import 'package:dds/dap.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 
@@ -25,6 +26,42 @@ main() {
       final breakpointLine = lineWith(testFile, breakpointMarker);
 
       await client.hitBreakpoint(testFile, breakpointLine);
+    });
+
+    test('resolves and updates breakpoints', () async {
+      final client = dap.client;
+      final testFile = dap.createTestFile(simpleBreakpointResolutionProgram);
+      final setBreakpointLine = lineWith(testFile, breakpointMarker);
+      final expectedResolvedBreakpointLine = setBreakpointLine + 1;
+
+      // Collect any breakpoint changes during the run.
+      final breakpointChangesFuture = client.breakpointChangeEvents.toList();
+
+      Future<SetBreakpointsResponseBody> setBreakpointFuture;
+      await Future.wait([
+        client
+            .expectStop('breakpoint',
+                file: testFile, line: expectedResolvedBreakpointLine)
+            .then((_) => client.terminate()),
+        client.initialize(),
+        setBreakpointFuture = client.setBreakpoint(testFile, setBreakpointLine),
+        client.launch(testFile.path),
+      ], eagerError: true);
+
+      // The initial setBreakpointResponse should always return unverified
+      // because we verify using the BreakpointAdded/BreakpointResolved events.
+      final setBreakpointResponse = await setBreakpointFuture;
+      expect(setBreakpointResponse.breakpoints, hasLength(1));
+      final setBreakpoint = setBreakpointResponse.breakpoints.single;
+      expect(setBreakpoint.verified, isFalse);
+
+      // The last breakpoint change we had should be verified and also update
+      // the line to [expectedResolvedBreakpointLine] since the breakpoint was
+      // on a blank line.
+      final breakpointChanges = await breakpointChangesFuture;
+      final updatedBreakpoint = breakpointChanges.last.breakpoint;
+      expect(updatedBreakpoint.verified, isTrue);
+      expect(updatedBreakpoint.line, expectedResolvedBreakpointLine);
     });
 
     test('does not stop at a removed breakpoint', () async {