+++ pkg/dds/test/dap/integration/debug_variables_test.dart
@@ -17,7 +17,7 @@ main() {
   tearDown(() => dap.tearDown());
 
   group('debug mode variables', () {
-    test('provides variable list for frames', () async {
+    test('provides local variable list for frames', () async {
       final client = dap.client;
       final testFile = dap.createTestFile('''
 void main(List<String> args) {
@@ -53,6 +53,36 @@ void foo() {
         '''
             args: List (0 items), eval: args, 0 items
             myVariable: 1, eval: myVariable
+        ''',
+      );
+    });
+
+    test('provides global variable list for frames', () async {
+      final client = dap.client;
+      final testFile = dap.createTestFile('''
+final globalInt = 1;
+final globalString = 'TEST';
+final globalMyClass = MyClass();
+
+void main(List<String> args) {
+  globalMyClass;
+  print(''); $breakpointMarker
+}
+
+class MyClass {}
+    ''');
+      final breakpointLine = lineWith(testFile, breakpointMarker);
+
+      final stop = await client.hitBreakpoint(testFile, breakpointLine);
+      final topFrameId = await client.getTopFrameId(stop.threadId!);
+
+      await client.expectScopeVariables(
+        topFrameId,
+        'Globals',
+        '''
+            globalInt: 1, eval: globalInt
+            globalMyClass: MyClass, eval: globalMyClass
+            globalString: "TEST", eval: globalString
         ''',
       );
     });